refactor(core): 將 utils.py 重構為模組化套件
將單一 core/utils.py 檔案拆分為多個職責明確的模組，改善程式碼可維護性與可讀性。

主要變更：
- 建立 core/utils/ 套件結構，包含四個子模組：
  - logging.py: Logger 設定與 TqdmLogSink
  - filesystem.py: 檔案系統操作工具
  - alternate_names.py: 替代名稱處理
  - geoname_ids.py: GeoName ID 計算工具
- 統一匯出介面於 __init__.py，保持向後相容性
- 移除所有模組中的 sys.path.append hack，改用相對匯入
- 更新 10 個檔案的匯入路徑以使用新模組結構
- 為所有新模組添加完整的 Google 格式 docstring

通過檢查：
- ✅ ruff format 與 check
- ✅ 所有功能測試正常運作
- ✅ main.py 各子命令運作正常
